# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for Nihon Vocabulary Application",
        "title": "Nihon Vocabulary",
        "contact": {
            "name": "Songvut Nakrong",
            "email": "songvut.nakrong@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/vocabulary": {
            "post": {
                "description": "Create Vocabulary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabulary"
                ],
                "summary": "Create Vocabulary",
                "parameters": [
                    {
                        "description": "Vocabulary ID",
                        "name": "vocabularyInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VocabularyInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/vocabulary/": {
            "get": {
                "description": "Get Vocabularies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabulary"
                ],
                "summary": "Get Vocabularies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/vocabulary/{id}": {
            "get": {
                "description": "Get Vocabulary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabulary"
                ],
                "summary": "Get Vocabulary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Vocabulary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabulary"
                ],
                "summary": "Update Vocabulary",
                "parameters": [
                    {
                        "description": "Vocabulary ID",
                        "name": "vocabularyInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VocabularyInput"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Vocabulary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocabulary"
                ],
                "summary": "Delete Vocabulary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vocabulary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Japanese": {
            "type": "object",
            "properties": {
                "kana": {
                    "type": "string"
                },
                "kanji": {
                    "type": "string"
                },
                "romaji": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.VocabularyInput": {
            "type": "object",
            "properties": {
                "english": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "examples": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image": {
                    "type": "string"
                },
                "japanese": {
                    "$ref": "#/definitions/models.Japanese"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thai": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "vocie": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "ApiKeyAuth",
            "in": "header"
        }
    }
}